# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WeeklyRun API",
    "description": "WeeklyRun - AI Meeting Orchestration API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/transcription/meetings/{meeting_id}/transcribe": {
      "post": {
        "tags": [
          "transcription"
        ],
        "summary": "Start Transcription",
        "description": "Start transcription processing for a meeting.\n\nThis endpoint triggers STT processing in the background.\nAfter STT, runs chat filtering to remove casual talk (can be disabled).\nUse GET /meetings/{meeting_id}/progress to poll for status.\n\nArgs:\n    meeting_id: Meeting UUID\n    request: Optional request body with enable_chat_filter flag",
        "operationId": "start_transcription_api_v1_transcription_meetings__meeting_id__transcribe_post",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TranscriptionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transcription/meetings/{meeting_id}/transcripts": {
      "get": {
        "tags": [
          "transcription"
        ],
        "summary": "Get Meeting Transcripts",
        "description": "Get all transcript segments for a meeting.",
        "operationId": "get_meeting_transcripts_api_v1_transcription_meetings__meeting_id__transcripts_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TranscriptSegmentResponse"
                  },
                  "title": "Response Get Meeting Transcripts Api V1 Transcription Meetings  Meeting Id  Transcripts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transcription/meetings/{meeting_id}/transcripts/text": {
      "get": {
        "tags": [
          "transcription"
        ],
        "summary": "Get Meeting Transcript Text",
        "description": "Get full transcript as plain text.\n\nArgs:\n    meeting_id: Meeting ID\n    include_speakers: Include speaker labels in output",
        "operationId": "get_meeting_transcript_text_api_v1_transcription_meetings__meeting_id__transcripts_text_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          },
          {
            "name": "include_speakers",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Speakers"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Meeting Transcript Text Api V1 Transcription Meetings  Meeting Id  Transcripts Text Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TranscriptSegmentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "start_time": {
            "type": "number",
            "title": "Start Time"
          },
          "end_time": {
            "type": "number",
            "title": "End Time"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "speaker_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Label"
          },
          "speaker_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Name"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "id",
          "start_time",
          "end_time",
          "text",
          "speaker_label",
          "speaker_name",
          "confidence"
        ],
        "title": "TranscriptSegmentResponse",
        "description": "Transcript segment response."
      },
      "TranscriptionRequest": {
        "properties": {
          "enable_chat_filter": {
            "type": "boolean",
            "title": "Enable Chat Filter",
            "default": true
          }
        },
        "type": "object",
        "title": "TranscriptionRequest",
        "description": "Request body for transcription."
      },
      "TranscriptionStatusResponse": {
        "properties": {
          "meeting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meeting Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "segments_count": {
            "type": "integer",
            "title": "Segments Count"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "meeting_id",
          "status",
          "segments_count",
          "duration_seconds",
          "error_message"
        ],
        "title": "TranscriptionStatusResponse",
        "description": "Transcription status response."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}